@@ -5,7 +5,7 @@ description = "Client for the Server-Sent Events protocol (aka EventSource)"
 repository = "https://github.com/launchdarkly/rust-eventsource-client"
 authors = ["LaunchDarkly"]
 edition = "2021"
-rust-version = "1.83.0"
+rust-version = "1.88.0"
 license = "Apache-2.0"
 keywords = ["launchdarkly", "feature-flags", "feature-toggles", "eventsource", "server-sent-events"]
 exclude = ["CHANGELOG.md"]
